Промежуточная корректировка проекта.

+++ src/main/java/com/example/projectadd/impl/AdsServiceImpl.java
@@ -4,7 +4,12 @@ import com.example.projectadd.DTO.AdsDTO;
 import com.example.projectadd.DTO.CreateAdsDTO;
 import com.example.projectadd.DTO.FullAdsDTO;
 import com.example.projectadd.DTO.ResponseWrapperAds;
+import com.example.projectadd.config.Role;
 import com.example.projectadd.exception.AdNotFoundException;
+import com.example.projectadd.exception.NoAccessException;
+import com.example.projectadd.model.Image;
+import com.example.projectadd.model.User;
+import com.example.projectadd.repository.CommentRepository;
 import com.example.projectadd.repository.mapper.AdsMapper;
 import com.example.projectadd.model.Ads;
 import com.example.projectadd.repository.AdsRepository;
@@ -12,7 +17,6 @@ import com.example.projectadd.service.AdsService;
 import com.example.projectadd.service.ImageService;
 import com.example.projectadd.service.UserService;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -26,13 +30,14 @@ import java.util.stream.Collectors;
 @Transactional
 public class AdsServiceImpl implements AdsService {
     private final AdsRepository repository;
+    private final CommentRepository commentRepository;
     private final AdsMapper adsMapper;
     private final ImageService imageService;
     private final UserService userService;
 
-
-    public AdsServiceImpl(AdsRepository repository, AdsMapper adsMapper, ImageService imageService, UserService userService) {
+    public AdsServiceImpl(AdsRepository repository, CommentRepository commentRepository, AdsMapper adsMapper, ImageService imageService, UserService userService) {
         this.repository = repository;
+        this.commentRepository = commentRepository;
         this.adsMapper = adsMapper;
         this.imageService = imageService;
         this.userService = userService;
@@ -44,29 +49,40 @@ public class AdsServiceImpl implements AdsService {
     }
 
     @Override
-    public AdsDTO update(int id, CreateAdsDTO createAds) {
-        Ads ads = repository.findById(id).orElseThrow(AdNotFoundException::new);
-        ads.setDescription(createAds.getDescription());
-        ads.setPrice(createAds.getPrice());
-        ads.setTitle(createAds.getTitle());
-        repository.save(ads);
-        return adsMapper.toAdsDto(ads);
+    public AdsDTO update(int id, CreateAdsDTO createAds, Authentication authentication) {
+        if (adAccessCheck(id, authentication)) {
+            Ads ads = repository.findById(id).orElseThrow(AdNotFoundException::new);
+            ads.setDescription(createAds.getDescription());
+            ads.setPrice(createAds.getPrice());
+            ads.setTitle(createAds.getTitle());
+            repository.save(ads);
+            return adsMapper.toAdsDto(ads);
+        }
+        throw new NoAccessException("No access to ad");
     }
 
-
-
-
     @Override
-    public AdsDTO updateImage(int id, MultipartFile file) {
-        return null;
+    public AdsDTO updateImage(int id, MultipartFile file, Authentication authentication) {
+        if (adAccessCheck(id, authentication)) {
+            Ads ads = repository.findById(id).orElseThrow(AdNotFoundException::new);
+            Image image = ads.getImage();
+            if (image != null) {
+                imageService.remove(image);
+            }
+            ads.setImage(imageService.uploadImage(file));
+            return adsMapper.toAdsDto(repository.save(ads));
+        }
+        throw new NoAccessException("No access to ad");
     }
 
-
-
-
     @Override
-    public void deleteById(int id) {
-        repository.deleteById(id);
+    public boolean deleteById(int id, Authentication authentication) {
+        if (adAccessCheck(id, authentication)) {
+            commentRepository.deleteAllByAds_Id(id);
+            repository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 
     @Override
@@ -75,7 +91,7 @@ public class AdsServiceImpl implements AdsService {
     }
 
     @Override
-    public AdsDTO adAd(CreateAdsDTO createAds, MultipartFile file, Authentication authentication) {
+    public AdsDTO addAd(CreateAdsDTO createAds, MultipartFile file, Authentication authentication) {
         Ads ads = new Ads();
         ads.setDescription(createAds.getDescription());
         ads.setPrice(createAds.getPrice());
@@ -108,5 +124,14 @@ public class AdsServiceImpl implements AdsService {
     @Override
     public FullAdsDTO getAdInfo(int id) {
         return adsMapper.toFullAdsDto(repository.findById(id).orElseThrow(AdNotFoundException::new));
+    }
+
+    private boolean adAccessCheck(int id, Authentication authentication) {
+        User user = userService.getUser(authentication.getName());
+        if (user.getAuthorities().contains(Role.ADMIN)) {
+            return true;
+        }
+        Ads ads = repository.getById(id);
+        return ads.getUser().getId() == user.getId();
     }
 }
